refactor: rename files and directories for clarity and consistency
- Renamed 'r&d/' to 'experimental/' for better purpose alignment.
- Renamed helper files for clarity:
  - 'store_new_product_details_helpers.py' → 'store_product_helpers.py'
  - 'existing_product_helpers.py' → 'product_update_helpers.py'
  - 'scraping_helpers.py' → 'scraper_helpers.py'
- Updated utility names for consistency:
  - 'my_logger.py' → 'logging_utils.py'.
  - 'scrapers/scrapers.py' → 'scraper_manager.py'.

Improves readability, consistency, and adherence to naming best practices.
