+++ app/services/helpers/product_update_helpers.py
