+++ experimental/celery/cloudamqp_config.py
