+++ experimental/celery/learning_celery.py
@@ -11,5 +11,5 @@ def add(x, y):
 
 
 # Run the Celery worker with the following command:
-# cd '.\r&d\celery\' && celery -A learning_celery worker --loglevel=info
-# celery -A ./r&d/celery/learning_celery worker --loglevel=info
+# cd '.\experimental\celery\' && celery -A learning_celery worker --loglevel=info
+# celery -A ./experimental/celery/learning_celery worker --loglevel=info
