+++ experimental/redis/test-redis.py
