+++ tests/manual_tests/insert_dummy_product.py
