+++ tests/manual_tests/test_jobs_crud.py
