+++ tests/manual_tests/test_products_crud.py
